vars: handle p cnf line
In case of empty formulas, or formulas where the largest variables
are missing, the solver would produce models that do not map those
variables. To mitigate, make sure the solver is aware of the p cnf
line of the input formula.

For formulas, where the number in the p cnf line is still too small,
the solver would still go forward, as those variables are handled
during parsing the corresponding clauses.