lsids: fix 'no-lsids' case
When using no-lsids, we expect the solver to behave as before the
changes of this series. However, when comparing the output of the
solver with deactivated LSIDS, changes in the search behavior have been
spotted. These changes are caused by modifications to the way variable
activities are bumped during conflict analysis.

This change reverts these changes in case LSIDS is not used.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>