solvers: add add-script
This script allows to add the source of MiniSat based solvers with a
relation to existing MiniSat solvers.

The script adds LINCENSE, README and relevant source and header files,
enforces the code style, and afterwards adds the changes with a commit,
referencing the given solver name. Finally, the created git object is
tagged with the solver name, where the tag is prefixed with 'solver-'.

Before the commit is created, compilation of the solver is tested.

Tagging is currently not done automatically, as there might be followup
changes to the created commit, before it is final, e.g. adding further
links and references to original work.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>