parallel: add solver class
A parallel solver needs to implement the basic interface of the solver
class. Furthermore, public members need to be available, so that the
solver can be used as a library as well.

This class introduces the minimal set of members and functions, so that
the solver can be used from the typical solver entry point main.

The current code supports only using a single solver. Function calls are
forwarded to this single solver.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>