build: allow to select build type
By default, MergeSat should be build as sequential solver with
simplification. However, we want to start using the parallel solver
as well. Hence, allow to switch the type of solver that is build
based on compile time arguments.

This change adds a make variable 'BUILD_TYPE', which allows to control
whether the parallel solver should be build. An example build
invocation is

  make d BUILD_TYPE=paallel

Note, when cleaning up build files, we also need to specify the correct
build type, so that we cleanup all files:

  make clean BUILD_TYPE=paallel

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>