search: explicitly update used assumptions
Until now, the caching of the number of previously used
assumptions was ineffective, essentially a no-op.

This change moves the behavior into a separate function,
and calls this function at the correct places.

While being at it, also add a return statement to the
search function - which should currently never be reached.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>