parallel: add pool object
For clause sharing, we need to implement an accessible clause pool. This
change adds such a structure. The structure itself is based on known
data structures.

To be able to use the structures more effectively, also add a
place-holder constructor for clauses, which allows to fill in the
values for the actual literals at a later point.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>